fix(db): Remove cross-join warning in treasury running totals update (#1728)
* fix(db): Remove cross-join warning in treasury running totals update

* (fix): Treasury ConversionRateZltoUsd precision (numeric(18,10))

* (fix): Use numeric division for treasury conversion rate seed

+++ src/api/src/infrastructure/Yoma.Core.Infrastructure.Database/Migrations/20260310033518_ApplicationDb_Referral_Program_And_Treasury_Seeding.cs
@@ -65,7 +65,7 @@ namespace Yoma.Core.Infrastructure.Database.Migrations
           20000,
           NULL,
           NULL,
-          1.0 / 45,
+          1::numeric / 45,
           now(),
           '8929632E-2911-42FF-9A44-055DEF231B87',
           now(),
